jtt: use jruby-truffle
