+++ bin/jruby-truffle-tool
@@ -3,4 +3,7 @@
 bin="$(cd "$(dirname "${BASH_SOURCE[0]}")" && pwd)"
 root="$(dirname "$bin")"
 
-exec $bin/jruby $root/lib/ruby/truffle/jruby-truffle-tool/bin/jruby-truffle-tool "$@"
+exec ${bin}/jruby-truffle \
+     -Xtruffle.backtraces.hide_core_files=false \
+     ${root}/lib/ruby/truffle/jruby-truffle-tool/bin/jruby-truffle-tool \
+     "$@"
