+++ lib/ruby/truffle/jruby-truffle-tool/lib/actionmailer_exclusions.yaml
@@ -1,0 +1,1 @@
+---
